UDP relay should batch messages
Implements some basic internal queueing/flushing logic using some
hardcoded constants for now.

General design:
- parse UDP messages as JSON (puke an error if invalid)
- add parsed Object to an Array queue
- when queue passes a threshold, flush it OR if an interval timer
  triggers flush it as well
- once flushed, reset the timer and clear the queue

This way we take an approach similar to my Java
code (https://github.com/voutilad/groom) where the data is both
windowed by time and quantity. Flushes happen when one of the two
thresholds are reached.

Also, this stubs out sending the Cypher-based format (for now) to the
topic:

{
  batch: [ { event1 }, { event 2 }, ... ],
  cypher: 'MERGE (something)'
}

The subscription listener (used for testing) is updated to expect the
format and report on the size of the detected batches in each message.